+++ CHANGELOG.md
@@ -37,6 +37,8 @@
   - Requires macOS 11.3 (20E232) or newer
 - Add offline TUI build
   - Allows for root patching without network connection
+- Add Legacy Wireless support for Monterey
+  - Applicable for BCM94328, BCM94322 and Atheros chipsets
 
 ## 0.2.4